+++ lib/cpalgo/algebra/bingcd.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <utility>
 
 
@@ -17,11 +18,11 @@
 // Verified:
 //  - https://onlinejudge.u-aizu.ac.jp/courses/lesson/1/ALDS1/1/ALDS1_1_B
 //
-inline unsigned long long gcd(unsigned long long a, unsigned long long b) {
+inline uint64_t gcd(uint64_t a, uint64_t b) {
     if (a == 0) return b;
     if (b == 0) return a;
 
-    unsigned long long g = 1;
+    uint64_t g = 1;
     while (b > 0) {
         if ((a & 1) == 0 && (b & 1) == 0) {
             g <<= 1;
